refactor(cli): move boot-source detection into its own module
`exec.rs` crossed the 1500-line production budget. Settling which source a
run boots from is a self-contained decision — it reads the flags, the working
directory and the runtime catalog and hands back one answer — so it moves
whole into `exec/detect.rs`: `Inference`, `ResolvedSource`, `resolve_run_source`
and the two helpers it calls, with the `source_resolution` and
`declared_binding_tests` suites that cover them.

`resolve_run_source` and `Inference` are re-exported from `exec`, so the
`machine run` call site is unchanged. exec.rs drops to 1334 production lines.
